+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### Added
 
+ - Added flag to set verbosity level.
+ - Validate TbProfiler schema version.
+ - Added CLI command for adding IGV annotation tracks
+
 ### Fixed
 
 ### Changed